feat: pretty messages
Relaying each message from `PlayerManager.broadcast()` turned out to be subpar because at that point every message had a player name prefixed to it and other bits of formatting baked in, resulting in undesirable output to the Discord channel.
On top of this, messages proxied from Discord to Minecraft would subsequently be relayed back to Discord as a server message because I was using `PlayerManager.broadcast()` to send relayed Discord messages -- sleep deprivation due to corporatism.

To solve both of these problems at the same time, I'm now using a bunch of more precise mixins to capture raw elements and format the final message how I'd like.